backend/requirements.txt
fastapi>=0.109.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
httpx>=0.26.0
python-multipart>=0.0.9
